requirements.txt
pandas==2.2.3
plotly==5.24.1
requests==2.32.3
streamlit==1.38.0
python-dotenv==1.0.1
matplotlib==3.9.2
seaborn==0.13.2
streamlit-aggrid==1.0.5


